+++ src/main/java/uk/co/duelmonster/minersadvantage/common/services/policy/PolicyCoreService.java
@@ -213,6 +213,7 @@ public final class PolicyCoreService {
         selected.ignoreIfValidTool(),
         selected.ignorePassiveMobs(),
         selected.blacklist(),
+        selected.blockBlacklist(),
         selected.selectionRules());
   }